+++ git_improved/commands/done.py
@@ -55,7 +55,7 @@ def merge_squash(merged_branch, message=None):
 
 def done_command():
     ensure_git_initialized()
-    ensure_working_tree_clean()
+    #ensure_working_tree_clean()
     ensure_branch_mergeable()
     
     current_branch = get_current_branch()
@@ -68,6 +68,9 @@ def done_command():
     # sort commits from the oldest to the most recent
     commits = list(reversed(get_current_branch_commits()))
 
+    # keep a track of original changelog
+    original_changelog = Changelog.parse('docs/changelog.md')
+
     # parse changelog and add branch changes to [unreleased] appropriate section
     changelog = Changelog.parse('docs/changelog.md')
     changelog.add_change(change_category, branch_description, [commit['message'] for commit in commits])
@@ -77,18 +80,29 @@ def done_command():
     subprocess.call(['code', 'docs/changelog.md'])
     input("Edit changelog if required. Then press [enter] to continue...")
 
-    # parse file to ensure syntax is valid
-    changelog = Changelog.parse('docs/changelog.md')
-    changelog.save('docs/changelog.md')
+    # parse changelog to ensure syntax is still valid after edition
+    edited_changelog = Changelog.parse('docs/changelog.md')
+    edited_changelog.save('docs/changelog.md')
 
+    # workout changes added to changelog for this branch
+    original_unreleased = original_changelog.get_unreleased()
+    edited_unreleased = edited_changelog.get_unreleased()
+
+    # Commit all changes
     subprocess.call(['git', 'add', '.'])
     subprocess.call(['git', 'commit', '-m', 'update changelog'])
 
     # workout squash commit message
     squash_description = category_icon + " " + branch_description
 
-    if len(commits) > 1:
-        squash_description +=  "\n" + "\n".join(["- %s"%commit['message'] for commit in commits])
+    changes_diff = edited_unreleased.difference(original_unreleased)
+    if len(changes_diff.sections) > 0:
+        changes = ["- %s"%change.description for change in changes_diff.sections[0].changes if change.indent > 0]
+    
+        # only add changes to description if more than one change was done
+        # to keep description more readable
+        if len(changes) > 1:
+            squash_description +=  "\n" + "\n".join(changes)
     
     # merge-squash all commits from current branch in main branch
     merge_squash(current_branch, message=squash_description)